Fix register writes/reads offsets

+++ VM/src/runtime/io.h
@@ -6,6 +6,6 @@
  */
 namespace fe::vm
 {
-	extern "C" uint64_t print(const uint64_t *, uint8_t);
-	extern "C" uint64_t println(const uint64_t *, uint8_t);
+	extern "C" uint64_t print(const uint8_t *, uint8_t);
+	extern "C" uint64_t println(const uint8_t *, uint8_t);
 } // namespace fe::vm
